+++ CHANGELOG.md
@@ -76,6 +76,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.34] - 2026-03-30
+
+### Test
+- Update tests/test_llx_mcp.py
+
+### Other
+- Update pyqual/integrations/llx_mcp.py
+- Update pyqual/integrations/llx_mcp_service.py
+
 ## [0.1.33] - 2026-03-30
 
 ### Other